backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
celery==5.4.0
redis==5.0.8
psycopg[binary]==3.2.1
dj-database-url==2.2.0
gunicorn==23.0.0
